requirements.txt
flask==3.1.0
requests==2.32.3
pyTelegramBotAPI==4.22.0
anthropic>=0.50.0
openai==1.57.0
PyGithub==2.4.0
pytz==2024.2
gunicorn==23.0.0
pdfplumber==0.11.4
psycopg2-binary==2.9.10
openpyxl==3.1.5
